SignatureX: move reading logic into SignatureX utility away from SerializationUtils.
The implementation is not yet ideal. There are trade-offs to be made.
One annoyance is that all fields of SignatureX are currently public.
This would imply that 'writeTo' method should be a utility method. After
further inspection it seems that we may want to make the signature
private. This is only used to verify against the expected signature.
Therefore one would not need direct access to the byte-array anyways. As
with the other cryptographic stuff, it may even be dangerous, as you
would be able to modify the source content.

To be continued ...