Converted Scalar type to perform operations on byte-arrays.  (Implementation needs improving)
This is a conversion with the intention to isolate the BigInteger
conversions in the Scalar-type operations. That way we can prepare as
much as possible for byte-array oriented scalar handling. An open FIXME
warns that Scalar needs further improvements.

In this commit, StateExpect{1,2,3}Test fail due to test data needing
updating. Investigation in progress ...