Point implementation modified to keep internal data as byte-array.
Note that the implementation of Point (and Scalar) are inefficient,
"bad" implementations. The decision was made to keep internal data in
byte-array representation such that migration to a more efficient
byte-array based elliptic curve implementation (and byte-array based
scalar operations) is simpler.
We have effectively isolated the "bad" implementation into the lines of
logic inside individual methods of the Point (or Scalar) class.