Introduced custom "keypair" classes for DSA and DH, stricter import-control.
Introduced import-control rules for banning use of 'java.security'
outside of package 'crypto'.

Introduced types DSAKeyPair and DHKeyPairJ as replacement for the
genericless generic java.security.KeyPair class, which seems to be more
of an implementation detail. Got rid of previously necessary type-casting.

Refactored all existing code to use DSAKeyPair and DHKeyPairJ types.
This is all OTRv3-related code, as there are different types for use
with OTRv4 logic.