Improved state transitioning, part 1: verify from-state before transitioning.
Ensure that we're operating on the right assumed "current state" before
transitioning. Bail out in case the assumption is wrong.

This is the first step in preparation of cleaning up state data after
transitioning. In later additions, we're going to clear the old state
immediately after transitioning to the new state.