Fine-tuned rotation and max_remote_i_seen variable use.
This is not exactly according to spec, however there is an open
discussion on how exactly this is supposed to work.
(https://github.com/otrv4/otrv4/pull/211)

The rationale is: given that the first set of sending/receiving keys is
generated based on the firstECDH and firstDH public keys, the party who
generates the receiving keys can assume that ratchet ID 0 is seen, i.e.
`max_remote_i_seen == 0`, due to the public keys already being known.

Update `max_remote_i_seen` on receiver key rotation, instead of when
processing first data message content. When processing data message
content, we already assume that the public keys are in place and the
`max_remote_i_seen` is up to date. This part does not exactly deviate
from spec, but rather the code structure is such that the logic is
separated over two functions.