Partial: message queueing - design issue to be resolved.
This is a partial implementation of message queueing. Currently the
newly added test fails.
The outstanding issue is that we need to decide based on what session
state we decide to send message as plaintext or queue the message. As a
session may consist of multiple instances, we need to know how to work
with all of these "slave" instances. A small markdown doc is written
to elaborate on the 2 issues that are - for now - identified.